+++ tools/pi/scripts/update.sh
@@ -1,0 +1,14 @@
+#!/usr/bin/bash
+source $(t3x -T)
+
+figlet "T 3 X   PI"
+slow_lol "Update & upgrade packages"
+sleep 1
+
+#verify_internet_access
+#apt_update
+#apt_has_updates
+#git_has_updates
+
+sudo_run "apt update"
+sudo_run "apt upgrade -y"